+++ commands/analyze/what-anime.js
@@ -15,7 +15,7 @@ module.exports = class WhatAnimeCommand extends Command {
 				{
 					key: 'screenshot',
 					prompt: 'What screenshot do you want to scan?',
-					type: 'image|avatar',
+					type: 'image',
 					default: msg => msg.author.displayAvatarURL({ format: 'png', size: 512 })
 				}
 			]
