+++ commands/image-edit/demotivational-poster.js
@@ -35,7 +35,7 @@ module.exports = class DemotivationalPosterCommand extends Command {
 				{
 					key: 'image',
 					prompt: 'What image would you like to edit?',
-					type: 'image|avatar',
+					type: 'image',
 					default: msg => msg.author.displayAvatarURL({ format: 'png', size: 1024 })
 				}
 			]
