+++ commands/image-edit/needs-more-jpeg.js
@@ -19,7 +19,7 @@ module.exports = class NeedsMoreJpegCommand extends Command {
 				{
 					key: 'image',
 					prompt: 'What image would you like to edit?',
-					type: 'image|avatar',
+					type: 'image',
 					default: msg => msg.author.displayAvatarURL({ format: 'png', size: 512 })
 				},
 				{
