+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "xiao",
-	"version": "90.1.0",
+	"version": "90.1.1",
 	"description": "Your personal server companion.",
 	"main": "Xiao.js",
 	"scripts": {
